+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# Upcoming
+
 # v1.4.2.2
 
 ## Fixes